Move table scan builder and optimize rule to read package
